+++ spec/renderer_spec.rb
@@ -11,6 +11,14 @@ RSpec.describe ApexCharts::Renderer do
       },
       div: {
         class: 'css-class'
+      },
+      dataLabels: {
+        formatter: {
+          function: {
+            args: "value",
+            body: 'return value + " rabbits";'
+          }
+        }
       }
     }
   }
@@ -23,6 +31,7 @@ RSpec.describe ApexCharts::Renderer do
       expect(parsed).not_to be_blank
       expect(parsed.css('div.css-class')).not_to be_empty
       expect(parsed.css('script')).not_to be_empty
+      expect(parsed.at("script:contains('function(value){return value + \\\" rabbits\\\"')")).not_to be_nil
     end
   end
 end